8263564: Consolidate POSIX code for runtime exit support: os::shutdown, os::abort and os::die
Backport-of: 554dd29fb67b9a25ad161920b57e238c06024938
